Modify default boostrap modal to allow page scrolling.
By default, when a Twitter Bootstrap modal pops up, the page is not scrollable because the modal has a fixed position. This is a problem when the modal is bigger than the screen height because the user can't see the elements on the modal that don't fit on the screen.